// apps/coinet-platform/src/l12/contracts/l12-constitutional-types.ts
export enum L12ForbiddenAction {
  PREDICTION_THEATER = 'PREDICTION_THEATER',
  CERTAINTY_CLAIM = 'CERTAINTY_CLAIM',
  FINAL_JUDGMENT_EMISSION = 'FINAL_JUDGMENT_EMISSION',
  RECOMMENDATION_EMISSION = 'RECOMMENDATION_EMISSION',
  TRADE_ACTION_EMISSION = 'TRADE_ACTION_EMISSION',
  SCENARIO_AS_GUARANTEE = 'SCENARIO_AS_GUARANTEE',
  SINGLE_PATH_FAKE_CERTAINTY = 'SINGLE_PATH_FAKE_CERTAINTY',
  INVALIDATION_OMISSION = 'INVALIDATION_OMISSION',
  TRIGGER_OMISSION = 'TRIGGER_OMISSION',
  CONDITION_OMISSION = 'CONDITION_OMISSION',
  LOWER_LAYER_TRUTH_REDEFINITION = 'LOWER_LAYER_TRUTH_REDEFINITION',
  VALIDATION_REBUILD = 'VALIDATION_REBUILD',
  REGIME_REBUILD = 'REGIME_REBUILD',
  SEQUENCE_REBUILD = 'SEQUENCE_REBUILD',
  HYPOTHESIS_REBUILD = 'HYPOTHESIS_REBUILD',
  SCORE_REBUILD = 'SCORE_REBUILD',
  L11_SCORE_VALUE_ONLY_CONSUMPTION = 'L11_SCORE_VALUE_ONLY_CONSUMPTION',
  L7_RESTRICTION_BYPASS = 'L7_RESTRICTION_BYPASS',
  L8_REGIME_POSTURE_IGNORE = 'L8_REGIME_POSTURE_IGNORE',
  L9_SEQUENCE_POSTURE_IGNORE = 'L9_SEQUENCE_POSTURE_IGNORE',
  L10_HYPOTHESIS_POSTURE_IGNORE = 'L10_HYPOTHESIS_POSTURE_IGNORE',
  L11_SCORE_RESTRICTION_IGNORE = 'L11_SCORE_RESTRICTION_IGNORE',
  CONTRADICTION_DOWNGRADE = 'CONTRADICTION_DOWNGRADE',
  ACTIVE_INVALIDATION_HIDE = 'ACTIVE_INVALIDATION_HIDE',
  MISSING_VISIBILITY_HIDE = 'MISSING_VISIBILITY_HIDE',
  DRIFT_STATUS_HIDE = 'DRIFT_STATUS_HIDE',
  PATH_CONFIDENCE_LAUNDERING = 'PATH_CONFIDENCE_LAUNDERING',
  SCENARIO_SPREAD_HIDE = 'SCENARIO_SPREAD_HIDE',
  RAW_STORAGE_BYPASS = 'RAW_STORAGE_BYPASS',
  L5_PERSISTENCE_BYPASS = 'L5_PERSISTENCE_BYPASS',
  L13_PLUS_CONSUMPTION = 'L13_PLUS_CONSUMPTION',
}
// ...
export enum L12OutputSurfaceClass {
  SCENARIO_SET = 'SCENARIO_SET',
  BASE_CASE_SCENARIO = 'BASE_CASE_SCENARIO',
  BULLISH_CONTINUATION_SCENARIO = 'BULLISH_CONTINUATION_SCENARIO',
  BEARISH_FAILURE_SCENARIO = 'BEARISH_FAILURE_SCENARIO',
  TRIGGER_PROFILE = 'TRIGGER_PROFILE',
  INVALIDATION_PROFILE = 'INVALIDATION_PROFILE',
  PATH_CONFIDENCE_PROFILE = 'PATH_CONFIDENCE_PROFILE',
  SCENARIO_SHIFT_CONDITION_SET = 'SCENARIO_SHIFT_CONDITION_SET',
  SCENARIO_RESTRICTION_PROFILE = 'SCENARIO_RESTRICTION_PROFILE',
  SCENARIO_EVIDENCE_READ_SURFACE = 'SCENARIO_EVIDENCE_READ_SURFACE',
  SCENARIO_LINEAGE_READ_SURFACE = 'SCENARIO_LINEAGE_READ_SURFACE',
}
// ...
export enum L12DownstreamConsumer {
  L13_JUDGMENT_LAYER = 'L13_JUDGMENT_LAYER',
  L14_DELIVERY_LAYER = 'L14_DELIVERY_LAYER',
  L15_FEEDBACK_LAYER = 'L15_FEEDBACK_LAYER',
  L16_GOVERNANCE_LAYER = 'L16_GOVERNANCE_LAYER',
  L17_PLUS_LAYER = 'L17_PLUS_LAYER',
}